update caculate score, using support card

+++ src/main/java/com/example/quizgame/dto/answer/AnswerResult.java
@@ -17,6 +17,6 @@ public class AnswerResult implements Serializable {
     private String selectedAnswer;
     private int score;
     private boolean correct;
-    private Float timeTaken;
+    private Long timeTaken;
     private String correctAnswer;
 }
